requirements.txt
filetype
pypdf[full]
Pillow
boto3

